some correction for the type or interface
